Refactor code and admin protect routes
