+++ server/assets/router.go
@@ -44,6 +44,12 @@ func Router(r *gin.Engine) {
 
 	}
 
+	// r.Static("/tmp", "./assets/tmp")
+	// r.GET("/gallery", imageListProvider)
+	// r.GET("/gallery/:id", imageDetailProvider)
+	// r.PUT("/gallery/:id", approveImage)
+	// r.DELETE("/gallery/:id", removeImage)
+
 	// Admin only routes
 	admin := r.Group("/")
 	admin.Use(middleware.UserAuthenticator, middleware.EmailVerified, middleware.AdminAuthenticator)
